+++ chalicelib/db/product/repository.py
@@ -1,0 +1,5 @@
+from chalicelib.interface.repository import Repository
+
+
+class ProductMongoRepository(Repository):
+    pass
